Reformatting the structure
